emp components
