+++ app/src/main/java/reservant_mobile/ui/components/Components.kt
@@ -136,7 +136,6 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.layout.ContentScale
 import androidx.compose.ui.layout.onGloballyPositioned
 import androidx.compose.ui.platform.LocalContext
-import androidx.compose.ui.platform.LocalDensity
 import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.text.SpanStyle
@@ -439,118 +438,6 @@ fun ShowErrorToast(context: Context, id: Int) {
     }
 }
 
-@Composable
-fun EmployeeCard(
-    employee: RestaurantEmployeeDTO,
-    onEditClick: () -> Unit,
-    onDeleteClick: () -> Unit
-) {
-    var showConfirmDeletePopup by remember { mutableStateOf(false) }
-
-    when {
-        showConfirmDeletePopup -> {
-            CountDownPopup(
-                icon = Icons.Filled.DeleteForever,
-                title = stringResource(id = R.string.confirm_delete_title),
-                text = stringResource(id = R.string.confirm_delete_text),
-                onConfirm = {
-                    onDeleteClick()
-                    showConfirmDeletePopup = false
-                },
-                onDismissRequest = { showConfirmDeletePopup = false },
-                confirmText = stringResource(id = R.string.label_yes_capital),
-                dismissText = stringResource(id = R.string.label_cancel)
-            )
-        }
-    }
-
-
-    Card(
-        modifier = Modifier
-            .fillMaxWidth()
-            .padding(8.dp),
-        elevation = CardDefaults.cardElevation(8.dp)
-    ) {
-        val localDensity = LocalDensity.current
-        var tabWidth by remember { mutableStateOf(0.dp) }
-        Box(
-            modifier = Modifier.fillMaxWidth(),
-            contentAlignment = Alignment.BottomEnd
-        ) {
-            Box(
-                modifier = Modifier.fillMaxWidth(),
-                contentAlignment = Alignment.TopStart
-            ) {
-
-                Column(
-                    modifier = Modifier.padding(16.dp)
-                ) {
-                    Text(
-                        text = "${employee.firstName} ${employee.lastName}",
-                        style = MaterialTheme.typography.titleMedium.copy(fontSize = 20.sp)
-                    )
-
-                    Text(
-                        text = "${stringResource(id = R.string.label_login_display)} ${employee.login}",
-                        style = MaterialTheme.typography.bodyMedium.copy(fontSize = 16.sp),
-                        modifier = Modifier.padding(top = 8.dp)
-                    )
-
-                    Text(
-                        text = "${stringResource(id = R.string.label_phone_display)} ${employee.phoneNumber}",
-                        style = MaterialTheme.typography.bodyMedium.copy(fontSize = 16.sp),
-                        modifier = Modifier
-                            .onGloballyPositioned { coordinates ->
-                                tabWidth = with(localDensity) { coordinates.size.width.toDp() }
-                            }
-                    )
-
-                    HorizontalDivider(
-                        modifier = Modifier
-                            .width(tabWidth),
-                        thickness = 1.dp,
-                        color = MaterialTheme.colorScheme.outline
-                    )
-
-
-                    if (employee.isHallEmployee) {
-                        Text(
-                            text = stringResource(id = R.string.label_employee_hall),
-                            style = MaterialTheme.typography.bodySmall
-                        )
-                    }
-                    if (employee.isBackdoorEmployee) {
-                        Text(
-                            text = stringResource(id = R.string.label_employee_backdoor),
-                            style = MaterialTheme.typography.bodySmall
-                        )
-                    }
-                }
-            }
-            Row {
-                val buttonModifier = Modifier
-                    .size(50.dp)
-                    .padding(6.dp)
-
-                SecondaryButton(
-                    modifier = buttonModifier,
-                    onClick = onEditClick,
-                    imageVector = Icons.Filled.Edit,
-                    contentDescription = "EditEmployee"
-                )
-
-                SecondaryButton(
-                    modifier = buttonModifier,
-                    onClick = {showConfirmDeletePopup = true},
-                    imageVector = Icons.Filled.DeleteForever,
-                    contentDescription = "DeleteEmployee"
-                )
-            }
-        }
-
-    }
-}
-
 
 @Composable
 fun BottomNavigation(
@@ -615,7 +502,7 @@ fun BottomNavigation(
 fun Heading() {
     val drawerState = rememberDrawerState(initialValue = DrawerValue.Closed)
     val scope = rememberCoroutineScope()
-
+    
     ModalNavigationDrawer(
         drawerState = drawerState,
         drawerContent = {
