+++ app/src/main/java/reservant_mobile/ui/components/EmployeeComponents.kt
@@ -1,0 +1,143 @@
+package reservant_mobile.ui.components
+
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.DeleteForever
+import androidx.compose.material.icons.filled.Edit
+import androidx.compose.material3.Card
+import androidx.compose.material3.CardDefaults
+import androidx.compose.material3.HorizontalDivider
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.mutableStateOf
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.layout.onGloballyPositioned
+import androidx.compose.ui.platform.LocalDensity
+import androidx.compose.ui.res.stringResource
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.reservant_mobile.R
+import reservant_mobile.data.models.dtos.RestaurantEmployeeDTO
+
+@Composable
+fun EmployeeCard(
+    employee: RestaurantEmployeeDTO,
+    onEditClick: () -> Unit,
+    onDeleteClick: () -> Unit
+) {
+    var showConfirmDeletePopup by remember { mutableStateOf(false) }
+
+    when {
+        showConfirmDeletePopup -> {
+            CountDownPopup(
+                icon = Icons.Filled.DeleteForever,
+                title = stringResource(id = R.string.confirm_delete_title),
+                text = stringResource(id = R.string.confirm_delete_text),
+                onConfirm = {
+                    onDeleteClick()
+                    showConfirmDeletePopup = false
+                },
+                onDismissRequest = { showConfirmDeletePopup = false },
+                confirmText = stringResource(id = R.string.label_yes_capital),
+                dismissText = stringResource(id = R.string.label_cancel)
+            )
+        }
+    }
+
+
+    Card(
+        modifier = Modifier
+            .fillMaxWidth()
+            .padding(8.dp),
+        elevation = CardDefaults.cardElevation(8.dp)
+    ) {
+        val localDensity = LocalDensity.current
+        var tabWidth by remember { mutableStateOf(0.dp) }
+        Box(
+            modifier = Modifier.fillMaxWidth(),
+            contentAlignment = Alignment.BottomEnd
+        ) {
+            Box(
+                modifier = Modifier.fillMaxWidth(),
+                contentAlignment = Alignment.TopStart
+            ) {
+
+                Column(
+                    modifier = Modifier.padding(16.dp)
+                ) {
+                    Text(
+                        text = "${employee.firstName} ${employee.lastName}",
+                        style = MaterialTheme.typography.titleMedium.copy(fontSize = 20.sp)
+                    )
+
+                    Text(
+                        text = "${stringResource(id = R.string.label_login_display)} ${employee.login}",
+                        style = MaterialTheme.typography.bodyMedium.copy(fontSize = 16.sp),
+                        modifier = Modifier.padding(top = 8.dp)
+                    )
+
+                    Text(
+                        text = "${stringResource(id = R.string.label_phone_display)} ${employee.phoneNumber}",
+                        style = MaterialTheme.typography.bodyMedium.copy(fontSize = 16.sp),
+                        modifier = Modifier
+                            .onGloballyPositioned { coordinates ->
+                                tabWidth = with(localDensity) { coordinates.size.width.toDp() }
+                            }
+                    )
+
+                    HorizontalDivider(
+                        modifier = Modifier
+                            .width(tabWidth),
+                        thickness = 1.dp,
+                        color = MaterialTheme.colorScheme.outline
+                    )
+
+
+                    if (employee.isHallEmployee) {
+                        Text(
+                            text = stringResource(id = R.string.label_employee_hall),
+                            style = MaterialTheme.typography.bodySmall
+                        )
+                    }
+                    if (employee.isBackdoorEmployee) {
+                        Text(
+                            text = stringResource(id = R.string.label_employee_backdoor),
+                            style = MaterialTheme.typography.bodySmall
+                        )
+                    }
+                }
+            }
+            Row {
+                val buttonModifier = Modifier
+                    .size(50.dp)
+                    .padding(6.dp)
+
+                SecondaryButton(
+                    modifier = buttonModifier,
+                    onClick = onEditClick,
+                    imageVector = Icons.Filled.Edit,
+                    contentDescription = "EditEmployee"
+                )
+
+                SecondaryButton(
+                    modifier = buttonModifier,
+                    onClick = { showConfirmDeletePopup = true },
+                    imageVector = Icons.Filled.DeleteForever,
+                    contentDescription = "DeleteEmployee"
+                )
+            }
+        }
+
+    }
+}
